Including editing comments in `create comment` test
